# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UPER IDP API Documentation",
        "description": "API documentation for UPER IDP SSO System (CENTRAL)",
        "contact": {
            "email": "burhan.mafazi@universitaspertamina.ac.id"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Get list of notifications for authenticated user",
                "operationId": "cac3bdde44f0be0512f7d05c0a6064e4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of notifications retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notifications retrieved successfully"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/notifications"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "GET"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2023-06-01 10:00:00"
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/NotificationResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Create a new notification",
                "operationId": "1fd84d2c66035574da4902f416cbd96a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "subject",
                                    "content"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string",
                                        "example": "info"
                                    },
                                    "subject": {
                                        "type": "string",
                                        "example": "New Message"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "You have a new message"
                                    },
                                    "detail_url": {
                                        "type": "string",
                                        "example": "/messages/123"
                                    },
                                    "app_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notification created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notification created successfully"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/notifications"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "POST"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2023-06-01 10:00:00"
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/NotificationResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/notifications/{uuid}": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark a notification as read",
                "operationId": "4b35b904080730c2c52558df5478ca0e",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Notification UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification marked as read successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/notifications/53e6b0f8-4f7c-4a3a-b0f8-4f7c4a3ab0f8"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "PUT"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2023-06-01 10:00:00"
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notification marked as read successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Delete a notification",
                "operationId": "67c80aadf1f2eae6d9de3e9d4d05a9a0",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Notification UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/notifications/53e6b0f8-4f7c-4a3a-b0f8-4f7c4a3ab0f8"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "DELETE"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2023-06-01 10:00:00"
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notification deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/notifications/read-all": {
            "put": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Mark all notifications as read",
                "operationId": "b9d04806e753b10a003ef77094ff56a3",
                "responses": {
                    "200": {
                        "description": "All notifications marked as read successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/v1/notifications/read-all"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "PUT"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2023-06-01 10:00:00"
                                        },
                                        "total_data": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All notifications marked as read successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotificationResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "uuid": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "app_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "info"
                    },
                    "subject": {
                        "type": "string",
                        "example": "New Message"
                    },
                    "content": {
                        "type": "string",
                        "example": "You have a new message"
                    },
                    "detail_url": {
                        "type": "string",
                        "example": "/messages/123"
                    },
                    "is_read": {
                        "type": "boolean",
                        "example": false
                    },
                    "read_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "application": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "uuid": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "code": {
                                "type": "string"
                            },
                            "name": {
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer"
                            },
                            "uuid": {
                                "type": "string",
                                "format": "uuid"
                            },
                            "username": {
                                "type": "string"
                            },
                            "full_name": {
                                "type": "string"
                            },
                            "code": {
                                "type": "string"
                            },
                            "email": {
                                "type": "string",
                                "format": "email"
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with username and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}